// src/app/models/empleado.ts
export interface Empleado {
    id_emp?:       string;
    nom_emp:      string;
    ape_emp:      string;
    id_tip:        number;
    num_doc_emp:  string;
    direcc_emp:   string;
    telefono_emp: string;
    email:         string;
    pass_emp:     string;
    id_rol:        number;
    estado_emp:   number;
    nombre_rol?:   string;
    rec_pass?  :   string;
}